fixup! Incomplete JWT setup
LDAP is "correctly" doing a connect and a bind, but... it always returns True and conn.response is None.
This is the case regardless of whether my bind_dn and bind_password are correct, or gibberish.
Maybe ldap3.Connection() defaults to SASL socket auth instead of bind auth?

Also, the /login_test page is not working, I just get HTTP 401 back when using /docs test URL.
It ought to be giving me an auth popup, right?
Oh, maybe I have to manually click "Authorize" button in top-right.

Also looks like I need to manually write an HTML login popup, and
I guess have 401 Unauthorized pages show it as the body or something?