fixup! Incomplete JWT setup
Also, the /login_test page is not working, I just get HTTP 401 back when using /docs test URL.
It ought to be giving me an auth popup, right?
Oh, maybe I have to manually click "Authorize" button in top-right.

Fixed.  The login route was "/login/" but the go-here-to-log-in url was "/login" (no trailing slash).